pyproject.toml
[project]
name = "oc-p12-extraction-donnees-multimodales"
version = "0.1.0"
description = "Extraction de données multimodales pour la détection de fake news"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32",
    "feedparser>=6.0",
    "pandas>=2.2",
    "datasets>=3.0",
    "Pillow>=11.0",
    "tqdm>=4.67",
    "python-dotenv>=1.0",
    "pyarrow>=23.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "responses>=0.25",
]
